WIP
Issues:
Fixes #<issueid>

Problem:

Analysis:

Tests:

+++ test/functional/neutronless/test_disconnected_service.py
@@ -45,12 +45,34 @@ SERVICELIBDIR = os.path.join(DISTRIBUTIONROOT,
                              'service_library')
 CREATELB_SVC =\
     json.load(open(os.path.join(SERVICELIBDIR, 'createlb.json'), 'r'))
-UPDATELB_SVC =\
-    json.load(open(os.path.join(SERVICELIBDIR, 'updatelb.json'), 'r'))
+DISCONNECTED_SVC =\
+    json.load(open(os.path.join(
+        SERVICELIBDIR, 'disconnected.json'), 'r'))['service']
+CONNECTED_SVC =\
+    json.load(open(os.path.join(
+        SERVICELIBDIR, 'connected.json'), 'r'))['service']
 DELETELB_SVC =\
     json.load(open(os.path.join(SERVICELIBDIR, 'deletelb.json'), 'r'))
 
+for listener in DISCONNECTED_SVC['listeners']:
+    listener['provisioning_status'] = 'PENDING_CREATE'
+for pool in DISCONNECTED_SVC['pools']:
+    pool['provisioning_status'] = 'PENDING_CREATE'
+for monitor in DISCONNECTED_SVC['healthmonitors']:
+    monitor['provisioning_status'] = 'PENDING_CREATE'
+for member in DISCONNECTED_SVC['members']:
+    member['provisioning_status'] = 'PENDING_CREATE'
+    member['network_id'] = u'a8f301b2-b7b9-404a-a746-53c442fa1a4f'
 
+for listener in CONNECTED_SVC['listeners']:
+    listener['provisioning_status'] = 'PENDING_CREATE'
+for pool in CONNECTED_SVC['pools']:
+    pool['provisioning_status'] = 'PENDING_CREATE'
+for monitor in CONNECTED_SVC['healthmonitors']:
+    monitor['provisioning_status'] = 'PENDING_CREATE'
+for member in CONNECTED_SVC['members']:
+    member['provisioning_status'] = 'PENDING_CREATE'
+    member['network_id'] = u'a8f301b2-b7b9-404a-a746-53c442fa1a4f'
 
 @pytest.fixture(scope='module')
 def setup_registry_snapshot(bigip):
@@ -74,10 +96,11 @@ def setup_neutronless(request, bigip, setup_registry_snapshot):
         ordering = {'member': 1,
                     '/mgmt/tm/ltm/pool': 2,
                     'monitor': 3,
-                    '/mgmt/tm/ltm/virtual': 4,
-                    '/mgmt/tm/net/fdb/tunnel': 5,
-                    'mgmt/tm/net/tunnels/tunnel/': 6,
-                    '/mgmt/tm/sys/folder': 7}
+                    'virtual-address': 4,
+                    '/mgmt/tm/ltm/virtual': 5,
+                    '/mgmt/tm/net/fdb/tunnel': 6,
+                    'mgmt/tm/net/tunnels/tunnel/': 7,
+                    '/mgmt/tm/sys/folder': 8}
         def order_key(item):
             for k in ordering:
                 if k in item:
@@ -88,11 +111,10 @@ def setup_neutronless(request, bigip, setup_registry_snapshot):
     def remove_test_created_elements():
         posttest_registry = register_device(bigip)
         created = frozenset(posttest_registry) - pretest_snapshot
-        pp('created')
-        pp(created)
         ordered = _deletion_order(created)
         for selfLink in ordered:
-            posttest_registry[selfLink].delete() 
+            if 'virtual-address' not in selfLink:
+                posttest_registry[selfLink].delete() 
 
     request.addfinalizer(remove_test_created_elements)
     return bigip, wrappedicontroldriver, pretest_snapshot
@@ -110,123 +132,36 @@ def test_loadbalancer_CD(setup_neutronless):
         assert sf.name == '/' or sf.name == 'Common'
 
     # invoke CREATELB_SVC operation
-    wicontrold._common_service_handler(CREATELB_SVC)
+    pp(DISCONNECTED_SVC)
+    # pp(CREATELB_SVC)
+    wicontrold._common_service_handler(DISCONNECTED_SVC)
 
     # verify CREATELB_SVCd state
     active_folders = bigip.tm.sys.folders.get_collection()
+    vs = bigip.tm.ltm.virtuals.get_collection()[0]
+    # pp(vs.raw)
+    start_keys = frozenset(vs.__dict__)
+    assert vs.selfLink == 'https://localhost/mgmt/tm/ltm/virtual'\
+        '/~TEST_cd6a91ccb44945129ac78e7c992655eb~listener2?ver=11.6.0'
+    #wicontrold._common_service_handler(CONNECTED_SVC)
+    vsend = bigip.tm.ltm.virtuals.get_collection()[0]
+    # pp(vsend.raw)
+    stop_keys = frozenset(vsend.__dict__)
+    pp(stop_keys - start_keys)
+    pp(start_keys - stop_keys)
+    #for k in start_keys:
     assert len(active_folders) == 3
     folder_names = [sf.name for sf in active_folders]
     thirty_two_zeros = '0'*32
-    assert '/' in folder_names and\
-           'Common' in folder_names and\
-           'TEST_' + thirty_two_zeros in folder_names
+    # assert '/' in folder_names and\
+    #       'Common' in folder_names and\
+    #       'TEST_' + thirty_two_zeros in folder_names
 
     # invoke DELETELB_SVC XXX operation FAILS silently!
-    wicontrold._common_service_handler(DELETELB_SVC)
+    # wicontrold._common_service_handler(DELETELB_SVC)
 
     # verify DELETELB_SVCd state
-    end_folders = bigip.tm.sys.folders.get_collection()
-    assert len(end_folders) == 2
-    for sf in end_folders:
-        assert sf.name == '/' or sf.name == 'Common'
-
-
-listener_config = {'name': 'test_listener',
-                   'loadbalancer_id': CREATELB_SVC['loadbalancer']['id'],
-                   'protocol': 'HTTP',
-                   'protocol_port': 80,
-                   'provisioning_status': 'PENDING_CREATE',
-                   'id': 'b'*32}
-CREATE_LISTEN_SVC = copy.deepcopy(CREATELB_SVC)
-CREATE_LISTEN_SVC['listeners'].append(listener_config)
-DELETE_LISTEN_SVC = copy.deepcopy(CREATE_LISTEN_SVC)
-DELETE_LISTEN_SVC['listeners'][0]['provisioning_status'] = 'PENDING_DELETE'
-
-
-def test_listener_CD(setup_neutronless):
-    bigip, wicontrold, _ = setup_neutronless
-    assert not bigip.tm.ltm.virtuals.get_collection()
-    wicontrold._common_service_handler(CREATE_LISTEN_SVC)
-    active_virts = bigip.tm.ltm.virtuals.get_collection()
-    assert active_virts[0].name == 'test_listener'
-    wicontrold._common_service_handler(DELETE_LISTEN_SVC)
-    assert not bigip.tm.ltm.virtuals.get_collection()
-
-
-# pp(CREATE_LISTEN_SVC)
-pool_config = {'name': 'test_pool_anur23rgg',
-               'lb_algorithm': 'ROUND_ROBIN',
-               'listener_id': CREATE_LISTEN_SVC['listeners'][0]['id'],
-               'protocol': 'HTTP',
-               'provisioning_status': 'PENDING_CREATE',
-               'id': 'c'*32}
-
-CREATE_POOL_SVC = copy.deepcopy(CREATE_LISTEN_SVC)
-CREATE_POOL_SVC['pools'].append(pool_config)
-DELETE_POOL_SVC = copy.deepcopy(CREATE_POOL_SVC)
-DELETE_POOL_SVC['pools'][0]['provisioning_status'] = 'PENDING_DELETE'
-
-def test_pool_CD(setup_neutronless):
-    bigip, wicontrold, _ = setup_neutronless
-    assert not bigip.tm.ltm.pools.get_collection()
-    wicontrold._common_service_handler(CREATE_POOL_SVC)
-    assert bigip.tm.ltm.pools.get_collection()[0].name == 'test_pool_anur23rgg'
-    wicontrold._common_service_handler(DELETE_POOL_SVC)
-    assert not bigip.tm.ltm.pools.get_collection()
-
-
-member_config = {'subnet_id': CREATELB_SVC['loadbalancer']['vip_subnet_id'],
-                 'network_id': CREATELB_SVC['loadbalancer']['network_id'],
-                 'provisioning_status': 'PENDING_CREATE',
-                 'pool_id': CREATE_POOL_SVC['pools'][0]['id'],
-                 'protocol_port': 80,
-                 'id': 'd'*32}
-
-CREATE_MEMBER_SVC = copy.deepcopy(CREATE_POOL_SVC)
-CREATE_MEMBER_SVC['members'].append(member_config)
-DELETE_MEMBER_SVC = copy.deepcopy(CREATE_MEMBER_SVC)
-DELETE_MEMBER_SVC['members'][0]['provisioning_status'] = 'PENDING_DELETE'
-
-
-def test_member_CD(setup_neutronless):
-    bigip, wicontrold, _ = setup_neutronless
-    wicontrold._common_service_handler(CREATE_POOL_SVC)
-    assert bigip.tm.ltm.pools.get_collection()[0].name == 'test_pool_anur23rgg'
-    members = bigip.tm.ltm.pools.get_collection()[0].members_s
-    assert not members.get_collection()
-    wicontrold._common_service_handler(CREATE_MEMBER_SVC)
-    # SANITY CHECK assert bigip.tm.ltm.pools.get_collection()[0].name == 'test_pool_anur23rgg'
-    assert members.get_collection()
-    wicontrold._common_service_handler(DELETE_MEMBER_SVC)
-    assert not members.get_collection()
-
-
-monitor_config =  {'delay': 3,
-                   'pool_id': CREATE_POOL_SVC['pools'][0]['id'],
-                   'type': 'HTTP',
-                   'timeout': 13,
-                   'max_retries': 7,
-                   'provisioning_status': 'PENDING_CREATE',
-                   'id': 'e'*32}
-
-CREATE_HM_SVC = copy.deepcopy(CREATE_MEMBER_SVC)
-CREATE_HM_SVC['healthmonitors'].append(monitor_config)
-DELETE_HM_SVC = copy.deepcopy(CREATE_HM_SVC)
-DELETE_HM_SVC['healthmonitors'][0]['provisioning_status'] == 'PENDING_DELETE'
-
-def test_healthmonitor_CD(setup_neutronless):
-    bigip, wicontrold, _ = setup_neutronless
-    https = bigip.tm.ltm.monitor.https
-    init_monitors = set([x.selfLink for x in https.get_collection()])
-    assert len(init_monitors) == 2
-    test_sl = 'https://localhost/mgmt/tm/ltm/monitor/http/'\
-              '~TEST_00000000000000000000000000000000'\
-              '~TEST_eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee?ver=11.6.0'
-    expected_monitors = init_monitors|set((test_sl,))
-    wicontrold._common_service_handler(CREATE_POOL_SVC)
-    wicontrold._common_service_handler(CREATE_HM_SVC)
-    observed_with_test_mon = set([x.selfLink for x in https.get_collection()])
-    assert observed_with_test_mon == expected_monitors
-    wicontrold._common_service_handler(DELETE_HM_SVC)
-    observed_post_delete = set([x.selfLink for x in https.get_collection()])
-    assert observed_post_delete == init_monitors
+    # end_folders = bigip.tm.sys.folders.get_collection()
+    # assert len(end_folders) == 2
+    # for sf in end_folders:
+    #     assert sf.name == '/' or sf.name == 'Common'
